refactor: replace NewGameModal and add dispatch event
