+++ src/components/Modals/Modals.module.scss
@@ -51,3 +51,13 @@
     color: $font-color;
   }
 }
+
+.toggle {
+  justify-content: center;
+
+  span {
+    font-size: 1.5rem;
+  }
+}
+
+
